refactor: unify image system with provider-agnostic `ShokoImage` model
Replaced the per-provider image models (`TMDB_Image`, `AniDB_Episode_PreferredImage`, `AniDB_Anime_PreferredImage`, `Image_Base`) with an unified `ShokoImage` and `ShokoImage_Entity` persistence, backed by cross-references supporting
multiple entities per image across any data source.

Key changes:
- Added `ShokoImage`/ShokoImage_Entity models with indexed repositories
- Updated `IImage` and `IWithImages`, added `IWithPrimaryImage`, `IWithBackdropImage`, `IWithLogoImage`, `ImageData` and `ImageUpdateData`, and removed `IWithPortraitImage` in the abstraction
- Upgraded the `IImageManager` in the abstraction and it's underlying implementation
- Replace per-provider download jobs with single DownloadImageJob
- Added `PurgeImageJob`, `PurgeOrphanedImagesJob`, and updated `ValidateAllImagesJob`
- Added `DataEntityType` enum for entity-scoped operations
- Added `KeyedEntityLockHelper` using `AsyncKeyedLock` for coordinated access
- Added new image settings, including `ImageTemplateUrlConfiguration` for remote sources
- Added TMDb image CDN URL override in `TMDBSettings`
- Extend `IAniDBService` with image-related and metadata purge methods
- Database migration converts legacy images to ShokoImage (DatabaseFixes)
- Remove ImageUtils, ImageExtensions, and all other legacy image handling

+++ Shoko.Server/API/v3/Helpers/WebUIFactory.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using Shoko.Abstractions.Metadata.Enums;
 using Shoko.Server.API.v3.Models.Common;
 using Shoko.Server.API.v3.Models.Shoko;
-using Shoko.Server.Extensions;
 using Shoko.Server.Models.AniDB;
 using Shoko.Server.Models.Shoko;
 using Shoko.Server.Server;
@@ -70,7 +69,7 @@ public class WebUIFactory
         {
             ID = group.AnimeGroupID,
             Type = anime.AnimeType.ToV3Dto(),
-            Rating = new Rating { Source = "AniDB", Value = anime.Rating, MaxValue = 1000, Votes = anime.VoteCount }
+            Rating = new Rating { Source = "AniDB", Value = anime.Rating, MaxValue = 1000, Votes = (uint)anime.VoteCount }
         };
         if (anime.AirDate is { } airDate && airDate != DateTime.MinValue)
             result.AirDate = airDate.ToDateOnly();
